Add php 8.1 and 8.2 to travis and moved tests
