+++ .travis.yml
@@ -3,6 +3,8 @@ language: php
 php:
   - '7.4'
   - '8.0'
+  - '8.1'
+  - '8.2'
 
 notifications:
   email:
